refactor: add race package and move classes to race package
