+++ src/main/kotlin/racingcar/domain/Referee.kt
@@ -1,5 +1,7 @@
 package racingcar.domain
 
+import racingcar.domain.race.RaceResult
+
 class Referee(private val raceResult: RaceResult) {
 
     fun determineWinner(): List<String> {
