+++ src/main/kotlin/racingcar/domain/race/RaceHistory.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.race
 
 class RaceHistory {
