+++ src/test/kotlin/racingcar/domain/CarsTest.kt
@@ -8,6 +8,7 @@ import racingcar.domain.car.Car
 import racingcar.domain.car.Name
 import racingcar.domain.car.Position
 import racingcar.domain.numberGenerator.FixedNumberGenerator
+import racingcar.domain.race.Cars
 
 class CarsTest {
